\name{reverseC-package}
\alias{reverseC-package}
\alias{reverseC-pkg}
\docType{package}
\title{
\packageTitle{reverseC}
}
\description{
\packageDescription{reverseC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{reverseC}
\packageIndices{reverseC}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{reverseC}

Maintainer: \packageMaintainer{reverseC}
}
\references{
%~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%~~ simple examples of the most important functions ~~
}
