\name{reverseDirection}
\alias{reverseDirection}
\title{reverseDirection}
\usage{
reverseDirection(n = 1000, nSNP = 10, MAF = rep(0.04, nSNP),
  gamma0 = 0, gammaX = rep(0.1, nSNP), varM = 1, beta0 = 0,
  betaM = seq(from = 0, to = 2, length.out = 4), varY = 1,
  nSim = 500, plot.pdf = T, plot.name = "reverseDirection.pdf",
  alpha_level = 0.05, SEED = 1)
}
\arguments{
\item{n}{n}

\item{nSNP}{nSNP}

\item{MAF}{MAF}

\item{gamma0}{gamma0}

\item{gammaX}{gammaX}

\item{varM}{varM}

\item{beta0}{beta0}

\item{betaM}{betaM}

\item{varY}{varY}

\item{nSim}{nSim}

\item{plot.pdf}{plot.pdf}

\item{plot.name}{plot.name}

\item{alpha_level}{alpha_level}

\item{SEED}{SEED}
}
\description{
placeholder_text
}
\author{
Sharon Lutz
}